requirements.txt
flask>=3.0.0
requests>=2.31.0
python-owasp-zap-v2.4>=0.0.21
jinja2>=3.1.2
weasyprint>=60.0
streamlit>=1.35.0
